update vtu_output tests
